feat(#116): add namespace to container list
Add tabulate to format tables.
Refactor `get_all_containers` to add namespace to tuple.
