+++ client/pyproject.toml
@@ -24,6 +24,7 @@ classifiers = [
 python = "^3.8"
 kubernetes = "^19.15.0"
 docker = "^5.0.3"
+tabulate = "^0.8.10"
 
 [tool.poetry.dev-dependencies]
 flake8-bugbear = "^22.1.11"
